Add ability to add open issues, assign users
Add a column of open issues. Drag an open issue to a milestone to
assign it to that milestone. Drag an issue back onto the open issues
column to remove it from a milestone.

Assign users by clicking the ugly assign button. You can then choose
a user to assign to the issue.