Separated milestone model from issue ordering
I moved the ordering of issues within a milestone into `MilestoneView`.
This seems to make more sense as the model is just for storing and
fetching data. We layer issue ordering on top of that, which belongs in
the view.

I also separated storing that data into another object. That will allow
me to more easily change the backend storage to use github instead of
parse for #4.