Debounce milestone priority saves so they occur after 10s
We need to debounce saves on milestones so we're not littering their
branches with checkins. It's easier to debounce the call to save all
milestones in `index.js` as opposed to the `save` method on each
milestone because we'd have to return/manage a promise if we did the
latter.