Styling for tags - first pass
Here’s a way we can display tags on githeartissues. For this first
pass, it’s a read-only approach to tags… so if we want to have
functionality like adding/removing tags, I might have to rethink this
UI.

I have it marked up so that the background color of the tag from github
is styled inline (I believe that’s how @benweatherman said we should do
this). If there are NO tags for an issue, the block shouldn’t even show
up in the dom.