Use a single file for saving
The old format of having an individual file for each milestone was OK
from an architecture standpoint, but gets kinda noisy if you change a
lot of milestones.

The new format saves everything in a single file. As a bonus, we'll
also store the issue text so it's easier to look at the saved file to
see what's been happening.

- Hijacked all the console functions to print the date/time
- Updated to swanky new lodash 3.0

Fixes #24